requirements.txt
alembic==1.12.0
aniso8601==10.0.0
axios==0.4.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
email-validator==2.1.0
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
mdurl==0.1.2
migrate==0.3.8
numpy==2.2.3
optional-django==0.3.0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
Pygments==2.19.1
pytest==7.4.2
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.1
react==4.3.0
requests==2.32.3
rich==14.0.0
six==1.17.0
SQLAlchemy==2.0.21
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==2.3.7
WTForms==3.2.1
